Support monitoring of lock status
Store creation time and staleness age in the semaphore so that we can
detect holders that might have been left hanging from a client crash or
similar.

Add `/lock-status` route for showing the current state of locks,
protected by Basic auth and configurable username/password. This makes it
easier to see the current state of things without using `etcdctl`.

Add ´/stale-locks` endpoint geared towards external monitoring that will
show if any locks are stale, as well as including the relevant holders
so that information can be present in the alarm output. This endpoint is
protected by the same Basic auth as `/lock-status`.