Introduce FleetLock group management API
Remove configuration of groups from the config file and instead store
this information in etcd3, manageable via a basic API. This way changes
can be performed without restarting all instances of the fleetlock
server.

In order to detect changes to the config in etcd3 we set up a watcher
that notifies the server if the config is changed. This then causes the
server to refresh its internal config automatically.

The API comes with Swagger UI documentation available at the `/swagger`
endpoint where you can also try out creating and deleting groups.