Make argon2 threads a fixed amount
It turns out the number of threads selected affects the resulting hash,
so the code as written would output different hashes based on the number
of CPUs which is not good when generating a key for encryption.

Set it to a fixed number (4) based on recommendations in RFC 9106.

While here cleanup comments and make it more clear what values are
selected when creating the hash.