Store permission passwords in hashed format
Since we have migrated from storing permission passwords as "strings in
config file" to "stored in etcd3" we should probably follow best
practices around storing passwords in databases.

Use argon2 to generate hashes for the passwords, and store these
instead. In order to be able to recreate the correct hash at a later
time we store all argon2 input settings next to the hash and use these
when handling authentication.

This has also required us to introduce an additional fleetlock config
struct, where the passwords have been converted from the cleartext
password data being sent to the server to the hashed format we want to
store in the backend.

While here make argon2 hashing used for getting a key for certmagic
encryption configurable, since otherwise we were unable to read certdata
written to etcd3 before the prior commit where we changed the argon2
threads setting to be a static value.