Add login cache for FleetLock authentication
Since we migrated to using argon2 hashes for FleetLock authentication,
and each request is checked against this hash since we keep no session
state, this means each request was pretty resource intensive to
authenticate.

To ease the burden of handling successful logins we introduce a cache,
that will hold a limited amount of successful authentications. The cache
is keyed by the following data:
permission key ("workers-*", "workers-worker1", etc) + existing hashed password + existing salt + cleartext password

The actual cache key string is a sha256 hash of the above data to get
some basic obfuscation.

With this change we also make the config reloader purge the login cache
so we would not accidentally authenticate a request based on stale cache
contents.