Support configuring the "exposed port" in config
When running the process in a container it is possible the client that
clients connect to is different than the port that our process actually
listens to.

Now you can configure what port is exposed to clients with
"exposed_port" in the [server] settings, and this will be used for
building URLs used by clients even if the process is listening on
another port.